feat(tests): fixed major bugs on the tests
A special thanks to drnic (https://github.com/drnic) for pointing out these problems. I've pushed this change to ensure that all of the tests run and I've updated the readme with details on how to use this in tests if the wrapper is not used as a wrapper around the whole test.